feat: add a sub-header for the menus on the transaction page
