+++ resources/views/livewire/pages/user/transactions/index.blade.php
@@ -183,7 +183,8 @@ new #[Layout('layouts.app')] class extends Component {
 
 <section x-data="{ detailSidebarOpen: false }" x-cloak>
     <!-- Main Content with Animated Margin -->
-    <div class="transition-all duration-300 ease-in-out" :class="{ 'md:mr-[23rem]': detailSidebarOpen }">
+    <div class="transition-all duration-300 ease-in-out"
+        :class="{ 'md:mr-[17rem] lg:mr-[23rem] xl:mr-[27rem] 2xl:mr-[41rem]': detailSidebarOpen }"">
         @livewire('pages.user.containers.main-header', ['component' => 'pages.user.transactions.header', 'header' => 'Transactions'])
 
         <div class="flex-1 overflow-y-auto md:pt-4 pt-4 px-6 bg-base-200">
@@ -271,17 +272,8 @@ new #[Layout('layouts.app')] class extends Component {
                     </ul>
                 @else
                     <div class="flex flex-col items-center justify-center p-10 bg-base-200/30 ">
-                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-                            stroke="currentColor" class="size-16 text-base-300 mb-3">
-                            <path stroke-linecap="round" stroke-linejoin="round"
-                                d="M12 9v3.75m9-.75a9 9 0 11-18 0 9 9 0 0118 0zm-9 3.75h.008v.008H12v-.008z" />
-                        </svg>
-                        <span class="text-base-content/60 text-lg font-medium mb-1">
-                            @if (!empty($filters['search']))
-                                No transactions found matching your search.
-                            @else
-                                No transactions found
-                            @endif
+                        <span class="text-base-content text-lg font-medium mb-1">
+                            😴 No transactions found
                         </span>
                         <p class="text-base-content/40 text-sm mb-4">
                             @if (
@@ -289,9 +281,6 @@ new #[Layout('layouts.app')] class extends Component {
                                     !empty($filters['types']) ||
                                     !empty($filters['account_id']) ||
                                     !empty($filters['date_mode']))
-                                Try adjusting your filters or search criteria
-                            @else
-                                Start adding your transactions to track your finances
                             @endif
                         </p>
                         @if (
@@ -313,7 +302,7 @@ new #[Layout('layouts.app')] class extends Component {
                                     stroke-width="1.5" stroke="currentColor" class="size-5 mr-1">
                                     <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
                                 </svg>
-                                Add Your First Transaction
+                                Add Transaction
                             </button>
                         @endif
                     </div>
